requirements.txt
appdirs==1.4.3
asgiref==3.2.7
beautifulsoup4==4.9.1
blinker==1.4
click==7.1.2
distlib==0.3.0
Django==3.0.5
django-cors-headers==3.3.0
django-enumfield==2.0.1
django-environ==0.4.5
djangorestframework==3.11.0
environ==1.0
filelock==3.0.12
Flask==1.1.2
Flask-Mail==0.9.1
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
Markdown==3.2.1
MarkupSafe==1.1.1
Pillow==7.1.2
psycopg2==2.8.5
python-dotenv==0.13.0
pytz==2019.3
six==1.14.0
soupsieve==2.0.1
sqlparse==0.3.1
Unidecode==1.1.1
urllib3==1.25.9
virtualenv==20.0.17
Werkzeug==1.0.1
zipp==3.1.0
